+++ idf_component_manager/service_details.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 ''' Helper function to init API client'''
 import os
+import warnings
 from collections import namedtuple
 
 from idf_component_manager.utils import print_info
@@ -43,8 +44,23 @@ def get_token(profile, token_required=True):  # type: (dict[str, str], bool) -> 
 
 def get_profile(config_path=None, profile_name=None):  # type: (str | None, str | None) -> dict[str, str]
     config = ConfigManager(path=config_path).load()
-    profile = config.profiles.get(profile_name, {})
-    return profile
+
+    profile_name_env_deprecated = os.getenv('IDF_COMPONENT_SERVICE_PROFILE')
+
+    if profile_name_env_deprecated:
+        warnings.warn(
+            'IDF_COMPONENT_SERVICE_PROFILE environment variable is deprecated. '
+            'Please use IDF_COMPONENT_REGISTRY_PROFILE instead',
+            category=DeprecationWarning)
+
+    profile_name_env = os.getenv('IDF_COMPONENT_REGISTRY_PROFILE')
+
+    if profile_name_env and profile_name_env_deprecated:
+        warnings.warn(
+            'Both IDF_COMPONENT_SERVICE_PROFILE and IDF_COMPONENT_REGISTRY_PROFILE '
+            'environment variables are defined. The value of IDF_COMPONENT_REGISTRY_PROFILE is used.')
+
+    return config.profiles.get(profile_name_env or profile_name_env_deprecated or profile_name, {})
 
 
 def service_details(