+++ idf_component_tools/config.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
+import warnings
 from io import open
 
 import yaml
@@ -11,7 +12,7 @@ from six import string_types
 from idf_component_tools.constants import COMPILED_URL_RE
 from idf_component_tools.errors import FatalError
 
-from .constants import DEFAULT_COMPONENT_SERVICE_URL, IDF_COMPONENT_STORAGE_URL
+from .constants import IDF_COMPONENT_REGISTRY_URL, IDF_COMPONENT_STORAGE_URL
 
 DEFAULT_CONFIG_DIR = os.path.join('~', '.espressif')
 CONFIG_DIR = os.environ.get('IDF_TOOLS_PATH') or os.path.expanduser(DEFAULT_CONFIG_DIR)
@@ -20,7 +21,7 @@ CONFIG_SCHEMA = Schema(
     {
         Optional('profiles'): {
             Or(*string_types): {
-                Optional('service_url'): Or('default', Regex(COMPILED_URL_RE)),
+                Optional('registry_url'): Or('default', Regex(COMPILED_URL_RE)),
                 Optional('default_namespace'): And(Or(*string_types), len),
                 Optional('api_token'): And(Or(*string_types), len)
             }
@@ -58,11 +59,11 @@ class ConfigManager(object):
     def load(self):  # type: () -> Config
         """Loads config from disk"""
         if not os.path.isfile(self.config_path):
-            return Config({}).validate()
+            return Config({})
 
         with open(self.config_path, mode='r', encoding='utf-8') as f:
             try:
-                return Config(yaml.safe_load(f.read()))
+                return Config(yaml.safe_load(f.read())).validate()
             except yaml.YAMLError:
                 raise ConfigError(
                     'Cannot parse config file. Please check that\n\t%s\nis valid YAML file\n' % self.config_path)
@@ -73,23 +74,47 @@ class ConfigManager(object):
             yaml.dump(data=dict(config.validate()), stream=f, encoding='utf-8', allow_unicode=True)
 
 
+def get_api_url(url):  # type: (str) -> str
+    url = url.rstrip('/')
+
+    if url.endswith('/api'):
+        return '{}/'.format(url)
+
+    return '{}/api/'.format(url)
+
+
 def component_registry_url(registry_profile=None):  # type: (dict[str, str] | None) -> tuple[str | None, str | None]
     """
-    Returns registry and static files URLs.
+    Returns registry API endpoint and static files URLs.
 
     Priorities:
     Environment variables > profile value in `idf_component_manager.yml` file > built-in default
+
+    If storage URL is configured it will always be used for downloads of components
     """
 
+    env_registry_url = os.getenv('IDF_COMPONENT_REGISTRY_URL')
+    env_registry_api_url = os.getenv('DEFAULT_COMPONENT_SERVICE_URL')
 
-    env_registry_url = os.getenv('DEFAULT_COMPONENT_SERVICE_URL')
+    if env_registry_api_url:
+        warnings.warn(
+            'DEFAULT_COMPONENT_SERVICE_URL environment variable pointing to the registy API is deprecated. '
+            'Set component registry URL to IDF_COMPONENT_REGISTRY_URL',
+            category=DeprecationWarning)
+
+    if env_registry_url and env_registry_api_url:
+        warnings.warn(
+            'Both DEFAULT_COMPONENT_SERVICE_URL and IDF_COMPONENT_REGISTRY_URL '
+            'environment variables are defined. The value of IDF_COMPONENT_REGISTRY_URL is used.')
+
+    if env_registry_url:
+        env_registry_api_url = get_api_url(env_registry_url)
+
     env_storage_url = os.getenv('IDF_COMPONENT_STORAGE_URL')
-    if env_registry_url or env_storage_url:
-        return env_registry_url, env_storage_url
 
-    env_registry_profile_name = os.getenv('IDF_COMPONENT_SERVICE_PROFILE')
-    if env_registry_profile_name:
-        registry_profile = ConfigManager().load().profiles.get(env_registry_profile_name, {})
+    if env_registry_api_url or env_storage_url:
+        return env_registry_api_url, env_storage_url
+
     if registry_profile is None:
         registry_profile = {}
 
@@ -99,7 +124,7 @@ def component_registry_url(registry_profile=None):  # type: (dict[str, str] | No
         storage_url = profile_storage_url
 
     registry_url = None
-    profile_registry_url = registry_profile.get('url')
+    profile_registry_url = registry_profile.get('registry_url')
     if profile_registry_url and profile_registry_url != 'default':
         registry_url = profile_registry_url
 
@@ -107,8 +132,9 @@ def component_registry_url(registry_profile=None):  # type: (dict[str, str] | No
         return None, storage_url
 
     if not registry_url:
-        registry_url = DEFAULT_COMPONENT_SERVICE_URL
-    if not storage_url:
-        storage_url = IDF_COMPONENT_STORAGE_URL
+        registry_url = IDF_COMPONENT_REGISTRY_URL
 
-    return registry_url, storage_url
+        if not storage_url:
+            storage_url = IDF_COMPONENT_STORAGE_URL
+
+    return get_api_url(registry_url), storage_url