+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add the `pre_release` flag for the dependency to control downloads of pre-release versions
 - Print a note with the list of alternative targets when the solver cannot find a suitable component version for the current target but there are some candidates for other targets.
 - Add the `examples` field for the custom examples paths
+- Add new environment variables `IDF_COMPONENT_REGISTRY_URL` and `IDF_COMPONENT_REGISTRY_PROFILE` for configuring the component manager
 
 ### Fixed
 
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix crash on use of local components and `override_path` for namespaced components
 - Mixing stdout and stderr of git command
 - Dump manifest file inconsistency with escaped dollar sign
+- Fix schema validation of the `idf_component_manager.yml` config file
 
 ## [1.1.4] 2022-07-04